fix: price Long -> Double 대대적 변경 수정

+++ backend/funco-follow/src/main/java/com/found_404/funco/feignClient/client/TradeServiceClient.java
@@ -19,7 +19,7 @@ public interface TradeServiceClient {
     CoinValuationResponse getCoinValuations(@RequestParam Long memberId);
 
     @GetMapping("/api/crypto")
-    Map<String, Long> getCryptoPrice(@RequestParam List<String> tickers);
+    Map<String, Double> getCryptoPrice(@RequestParam List<String> tickers);
 
     @GetMapping("/trade/holding/{ticker}")
     HoldingCoinResponse getHoldingCoin(@RequestParam Long memberId, @PathVariable String ticker);
