+++ backend/funco-follow/src/main/java/com/found_404/funco/follow/domain/FollowTrade.java
@@ -41,10 +41,10 @@ public class FollowTrade extends BaseEntity {
 	private Long orderCash;
 
 	@Comment("체결 가격")
-	private Long price;
+	private Double price;
 
 	@Builder
-	public FollowTrade(Follow follow, String ticker, TradeType tradeType, Double volume, Long orderCash, Long price) {
+	public FollowTrade(Follow follow, String ticker, TradeType tradeType, Double volume, Long orderCash, Double price) {
 		this.follow = follow;
 		this.ticker = ticker;
 		this.tradeType = tradeType;
