+++ backend/funco-follow/src/main/java/com/found_404/funco/follow/dto/Trade.java
@@ -9,6 +9,6 @@ public record Trade(
 	TradeType tradeType,
 	Double volume,
 	Long orderCash,
-	Long price
+	Double price
 ) {
 }
